+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat(catalog): add optional description field and update plugin version. (`615996f`)
 - feat(plugins): update plugin catalog with new screenshot handling and optional summary field. (`9e7e4cf`)
 
 ## 1.9.3 - 2026-07-01